requirement.txt
python3.6
numpy==1.16.4
scikit-learn==0.21.2
munkres==1.1.2 
matplotlib==3.1.1 
scipy==1.3.0
torch==1.1.0
torchvision==0.3.0
pillow==6.1.0